Fixed the download promise resolution logic
Previously the download promise was resolved on read stream "end" event. Naturally this event fires before write stream "finish" event which indicates that all the write data has been flushed to the underlying system. This caused the promise to be resolved **before** the last chunk was written to a file.